Fix delta code generation for nested structs shapes
When generating and compiling the lambda controller delta code, i noticed few compilation errors, moslty about type mismatches between values and function parameters.
This was due to a misplacement of the arugments passed to `code.compareSlice` and `code.compareMap` functions when they are called from `code.compareStruct`. This commit corrects the order/placement of arguments passed to the `compareSlice` and `compareMap` functions