handle renamed fields in SetResource generation
If fields were renamed in some operations, the code generated for
`SetResource()` was not properly looking up the renamed field. This
patch updates `SetResource()` to first check if the field has been
renamed for the referred Operation and if so, look up the CRDField
object in the CRD.SpecFields collection instead of the StatusFields
collection.