always sort additional printer columns
When returning AdditionalPrinterColumns for a CRD, we now always sort
the list of PrinterColumn structs by their Name. This prevents the
non-deterministic behaviour we've seen in Issue #720.

Also, the `pkg/generate:Generator.GetCRDs()` method was incorrectly only
adding PrinterColumns to a CRD when the field was specified as being
sourced from a different shape than the Create input or output shape (in
other words, only when the `pkg/generate/config.FieldConfig.From` field
was non-nil). This patch corrects this so that any field in a resource
may be marked as a printer column.

Fixes Issue #720