remove import tracking from TypeDef collection
This commit replaces the tracking of unique imported packages from the
processing of TypeDef structs by the code generator. Instead, we now
use the same technique to avoid import errors that we do in the `sdk.go`
code:

```go
import (
	ackv1alpha1 "github.com/aws-controllers-k8s/runtime/apis/core/v1alpha1"
	"github.com/aws/aws-sdk-go/aws"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
)

// Hack to avoid import errors during build...
var (
	_ = &metav1.Time{}
	_ = &aws.JSONValue{}
	_ = ackv1alpha1.AWSAccountID("")
)
```

The above is all the packages that are referenced by any type
definition's attribute fields that is collected by the code generator
during API inference.