implement SetSDK for Secret nested fields
Implements the `pkg/generate/code.SetSDK()` function for all nested
fields that are SecretKeyReference types. This involved passing a field
path down into the recursive `setSDKForXXX()` functions to keep track of
where we were in the processing of nested fields. We look up whether a
field is a SecretKeyReference by calling the
`pkg/model.CRD:IsSecretField()` method, which accepts a field path,
which is why we needed to pass this field path down into the recursive
code generators.

Issue aws-controllers-k8s/community#743