Patch the correct observed state after a `rm.ReadOne` call
For some controllers we noticed an incorrect reconciliation behaviour
when a custom hook code was returning a `RequeueNeededAfter` error.

More precisely `resourceManager.ReadOne` was not returning the correct
observed state when `sdkFind` returns with an error. This bug was causing
 the reconciler to patch to wrong state (`desird` instead `observed`).
This patch fixes this issue by passing the correct observed state to
`onError`.