Add a new operation config to override the `SetOutput` wrapper field
In some AWS APIs `Describe` operations output contains the
wanted information in a wrapper field like `TableDescription` and
`GlobalTableDescription` ... However in some other APIs for the same
operation type, the wanted information is wrapped in different nested
object, like `BackupDescription.BackupDetails` for example.

This patch adds a new operation configuration that instructs the
code-generator to use a specific field path, in order to find the
struct that contains the information we want to set the output from.