Mark spec as required, status.atProvider as optional
This is required to fix https://github.com/crossplane/provider-aws/issues/697

In the vast majority of resources a spec is required, so I'd like to mark it as
such. I believe that in the edge cases where the spec really contains only
optional fields it's possible to write `spec: {}`.

We prefer not to mark status fields as required, since resources will always
exist at some point without their status and we consider 'requireness' to be
more of a human-facing constraint than a software-facing constraint that our
managed resource controllers should be subject to.

Signed-off-by: Nic Cope <negz@rk0n.org>