Move "printer column" options under the same generator config struct
Initially we used to have one config field under "is_printable" under
`ackconfig.FieldConfig` to instruct the code generator to generate
kubebuilder marker comments in order to include the field in `kubectl
get` response.

Time passed and we are in a situation where want to tune more the way
we print these fields and extend the `kubebuilder:printcolumns` comment
markers to include options like `priority`, `order` and default columns.

This patch reworks the field configuration structure and prepare the
laying ground for future print config fields.

This patch will simplify the realization of the following issues:
* https://github.com/aws-controllers-k8s/community/issues/821
* https://github.com/aws-controllers-k8s/community/issues/822
* https://github.com/aws-controllers-k8s/community/issues/823