Fix required fields for renamed ReadOne input (#113)
Description of changes:
The generator was failing with a panic when it could not find a renamed field in the `ReadOne` describe/create operation while creating the `GoCodeRequiredFieldsMissingFromReadOneInput` template.

This pull request ensures the code attempts to rename the field, if specified in the config, before using the field name in any output.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.