handle renamed fields for resource identifiers
The code that outputs Go code that sets `Spec` or `Status` fields to a
resource identifier field was not accounting for cases where we renamed
the field. Such is the case for many of the RDS controller's resources,
since all the identifier fields for all resources are "de-stuttured" by
renaming, for example, `DBSubnetGroupName` to just `Name` or
`DBClusterParameterGroupName` to just `Name`.

This patch simply ensures we account for renamed fields in the Spec and
Status during calls to `pkg/generate/code.SetResourceIdentifiers` and
adds a unit test to ensure we don't regress in the future.

Issue aws-controllers-k8s/community#852