Support nested Ignore.FieldPaths
The current implementation of IgnoreFields did not support nested
shapes. For Ignore.FieldPaths: Foo.Bar.Baz, it would ignore the whole of
Bar.

With this change, it supports multiple levels of nested fields.

A practical example for this is ses CreateTemplateInput.Template.TemplateName.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.