Graceful cancellation for network dependent commands
`ack-generate` uses go-git to clone and fetch repository tags, these
actions can take time (more than a minute for cloning) and do not
tollerate interruption (ctrl+C). Forcing a command to exit when it's
running a `git.Clone` or a `git.FetchTags` cause the process to write
uncomplete data on disk, and therefore it is impossible to re-run
`ack-generate` without cleaning up the local cache.

This patch adds a gracefull cancellation system for `ack-generate`
subcommads. Which will allow users to safely ctrl+C and re-run their
commands without any errors.