Graceful cancellation for network dependent commands (#123)
Issue N/A

`ack-generate` uses go-git to clone and fetch repository tags, these
actions are time consuming (more than a minute for cloning) and do
not tolerate interruption (ctrl+C). Forcing a command to exit when it's
running a `git.Clone` or a `git.FetchTags` cause the process to write
incomplete data on disk,  therefore re-running `ack-generate` will
always return an error. The only solution for this problem is to delete
the local sdk cache and let `ack-generate` reclone (without interruption)

This patch adds a graceful cancellation system for `ack-generate`
subcommads. Which will allow users to safely ctrl+C and re-run their
commands without any errors.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.