Add service metadata configuration file (#126)
Description of changes:
Creates a new specification for a `metadata.yaml` file that resides in the root of the controller repository. The purpose of this file is to provide human-friendly display information about the service (long name, acronym, links) and to provide a source of truth for API versions.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.