Allow member fields from other operations in spec (#139)
Description of changes:
Allow the use of member fields contained within other API operations within a resource's spec. For example, `LoggingStatus` for S3 is only in the `PutBucketLogging` operation, but should be settable from spec.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.