Expose `SetSDKForStruct` and `SetResourceForStruct` in templates (#140)
Description of changes:
Makes `setSDKForStruct` and `setResourceForStruct` public, so that they are accessible for other parts of the generator. Exposes `SetSDKForStruct` as `GoCodeSetSDKForStruct` and `SetResourceForStruct` as `GoCodeSetResourceForStruct` in the generator, and includes `SDKAPI` as a template variable to allow querying of API shapes to use with parameters in the new template methods.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.