Generate typedefs for payload fields in `Spec` and `Status` (#146)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/864

Description of changes:
When checking payload type defs, if they are used in the spec or status are included, otherwise they are still excluded. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.