use reflect to compare map/slice struct
Although it's not what I wanted and isn't the most efficient thing we
could do, I'm implementing the slice of struct and map of struct
comparisons using `reflect.DeepEqual`. Long-term, I'd like to enable our
generator-config CompareConfig for nested struct fields, and to do that
we cannot use `reflect.DeepEqual`, however for the time-being, since
slice and map of struct is a hole in our delta logic, I'm pushing up
this workaround.