Allow override resource identifier to ARN (#163)
Closes https://github.com/aws-controllers-k8s/community/issues/909

Description of changes:
The SageMaker `ModelPackage` (in some cases) needs to be described using the resource ARN, even though the spec has a `ModelPackageName` and the `DescribeModelPackage` input shape has a `ModelPackageName` field. This pull requests allows the service teams to set the `primary_identifier_field_name` as `"ARN"`, to override the default generator logic in order to set the `ACKResourceMetadata.ARN` field to be `identifier.ARN`. 

Also refactored the identifier code to use the new `FindIdentifiersInShape` common method and to use the `setResourceForScalar` method.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.