Support nested paths in field config specifically list type members (#183)
Issue #, if available:
https://github.com/aws-controllers-k8s/community/issues/932

Description of changes:
`getMemberByPath` checks the shape type as it loops through members to ensure lists are handled correctly. Currently I have simply ignored the empty element created by the splitting of the string on the single dot instead of handling the double dot notation explicitly. Open to comments. 
Also added a Unit Test for the same. 

Testing
1. Generated code for ApplicationAutoscaling and ensured CreationTime field now works as expected. Linked changes [in this PR](https://github.com/aws-controllers-k8s/applicationautoscaling-controller/pull/39).
2. Ensured that the SageMaker Controller can be generated using new changes with no diff as expected. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.