Optionally log into ECR public (#189)
Description of changes:
Created a new helper method to determine whether the user is currently logged into the ECR public Docker registry. Only if they are not already logged in do we attempt to log in. This will allow the test infrastructure to log in once at the beginning (with a separate role) and ensure that login is stored even after we assume another role.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.