fixing formatting in container image to be consistent with image (#206)
Issue #, if available:
N/A

Description of changes:
This PR aims to keep how `containerImage` tag is generated/built consistent and insync with how `image` tag in the cluster service version is generated. This makes readability of both sections of code/templates/sh's/etc the same and easier in the future if a change in the formate is needed.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.