Export `GroupVersionResource` and `GroupKind` variables (#212)
Fixes https://github.com/aws-controllers-k8s/community/issues/998

Description of changes:
`GroupVersionResource` is used by dynamic clients to specify the CRD to be CRUD'd.
`GroupKind` is sometimes used in place where the version is not necessarily specified. I chose to export this mainly because it could be used down the line by other dynamic clients.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.