Rename ServiceID reference to ServiceIDClean (#214)
Issue #, if available:
aws-controllers-k8s/community#1009

Description of changes:
ServiceID was removed in a recent commit. One reference was left in clusterserviceversion.yaml.tpl. This PR simply renames that reference to ServiceIDClean to prevent breakage.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.