Consolidate service naming nomenclature (#213)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/994

Description of changes:
This pull request replaces all templates and variables with a defined nomenclature. The current code-generator uses service name terms interchangeably, which eventually lead to the linked issue. While PR #211 solved the logical issue of overriding the service model name, in some cases, it did not do a clear job of clarifying what names exist for any service and when they are applicable.

There are 3 different names that are used to programmatically reference a service. Here are the terms I am using in this PR to refer to them, and their given purpose:
- Service ID
  - The `metadata.serviceID` field from the model `api-2.json` file
- Service model name
  - The path to the model `api-2.json` file
- Service package name
  - The AWS SDK Go package path
  - Used by the ACK controller's for their name 

Some examples on when these differ:
- Step Functions uses `sfn` for its ID and alias, but `states` for its model name
- Elastic Load Balancing v2 uses `elbv2` for its alias, but `elasticloadbalancingv2` for its ID and model name

There are also 2 names defined in the controller metadata file that are not used in any parts of the code, but are used in documentation:
- Service name
  - The full name of the service as defined by the product's documentation page (eg. `Amazon Elastic Cloud Compute`)
-  Service short name
  - An informal, shortened name of the service (eg. `EC2`)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.