update olm script to remove AdoptedResource from the bundle that is generated (#224)
Issue #, if available:
aws-controllers-k8s/community#1006
This does not close this issue, this is just one piece needed. The runtime also needs to be update.

Description of changes:
This PR updates the `olm` scripts in a way that the `AdpotedResource` is no longer generated in the bundle that gets produced on disk. Since the OLM script works off a `temp` directory, this does not update/affect any of the artifacts on disk in the repo. The bundle that this generates has been tested against an OpenShift cluster vai OLM install. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.