Add `Dereference` template function (#230)
Description of changes:
This PR adds a Go template function for de-referencing strings. Go templates do not, out of the box, support string de-referencing when doing string comparison. When comparing a string pointer to a pointer it gives a fatal type mismatch error. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.