allow ignore nested fieldpaths (#260)
When implementing the update code paths for OpenSearchService, I found
out I needed to ignore the AutoTuneOptions.RollbackOnDisable nested
field within the UpdateDomainConfigInput shape. When I went to put the
following in the `generator.yaml` file:

```yaml
ignore:
  field_paths:
    - UpdateDomainConfigInput.AutoTuneOptions.RollbackOnDisable
```

What ended up happening is that the *entire* AutoTuneOptions field was
ignored instead of just the RollbackOnDisable field.

This patch fixes this situation by modifying the
`pkg/model.Model:ApplyIgnoreShapeRules()` method to account for nested
field paths and not just a two-level field path that has the Input or
Output shape name as the first part of the field path and the top-level
member field name as the second part of the field path.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.